changed spaceship motion
Now, move the mouse to change the rotation of the space ship. Only changes when the mouse is moving. Click and hold the mouse button to move forward in the direction of motion.
CHanged panel so that the ship can't move off the screen. The screen shifts (translate) so the ship stays on screen