add health
Replaced "progress" with health. Adding a health data item to SpaceObject and decrementing it when collisions occur. I overrode this for spacestations since they have a forcefield.
Now, if any space object gets too much damage it will be removed from the game. If it is the mainship then the game is over and the player gets a popup to that effect.
Also updated the help screen some.